+++ src/main/java/it/gov/pagopa/register/mapper/product/CookingHobsProductMapper.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.register.configuration.initiative.mapper;
+package it.gov.pagopa.register.mapper.product;
 
-import it.gov.pagopa.register.configuration.initiative.model.CategoryConfig;
+import it.gov.pagopa.register.model.initiative.CategoryConfig;
 import it.gov.pagopa.register.enums.ProductStatus;
 import it.gov.pagopa.register.model.operation.Product;
 import org.apache.commons.csv.CSVFormat;
@@ -20,8 +20,8 @@ import static it.gov.pagopa.register.mapper.operation.ProductMapper.limitName;
 import static it.gov.pagopa.register.mapper.operation.ProductMapper.normalizeCsvCode;
 import static it.gov.pagopa.register.utils.CsvUtils.DELIMITER;
 
-@Component("DECODER")
-public class DecoderProductMapper implements ProductMapperStrategy {
+@Component("COOKINGHOBS")
+public class CookingHobsProductMapper implements ProductMapperStrategy {
 
   @Override
   public String extractBusinessKey(CSVRecord record, CategoryConfig config) {
@@ -43,7 +43,7 @@ public class DecoderProductMapper implements ProductMapperStrategy {
     String codeProduct = normalizeCsvCode(csvRecord.get(CODE_PRODUCT));
     String gtinCode = normalizeCsvCode(csvRecord.get(CODE_GTIN_EAN));
 
-    String productName = CATEGORIES_TO_IT_S.get(category) + " " + csvRecord.get(BRAND) + " " + csvRecord.get(MODEL);
+    String productName = CATEGORIES_TO_IT_S.get(COOKINGHOBS) + " " + csvRecord.get(BRAND) + " " + csvRecord.get(MODEL);
     String fullProductName = gtinCode + " - " + productName;
 
     return Product.builder()
@@ -56,7 +56,8 @@ public class DecoderProductMapper implements ProductMapperStrategy {
       .productCode(codeProduct)
       .initiativeId(initiativeId)
       .gtinCode(gtinCode)
-      .category(category)
+      .category(COOKINGHOBS)
+      .countryOfProduction(csvRecord.get(COUNTRY_OF_PRODUCTION))
       .brand(csvRecord.get(BRAND))
       .model(csvRecord.get(MODEL))
       .capacity("N\\A")
@@ -82,6 +83,7 @@ public class DecoderProductMapper implements ProductMapperStrategy {
         product.getGtinCode(),
         product.getProductCode(),
         product.getCategory(),
+        product.getCountryOfProduction(),
         product.getModel(),
         product.getBrand()
       );
