+++ src/main/java/it/gov/pagopa/register/repository/initiative/InitiativeRepository.java
@@ -1,13 +1,8 @@
 package it.gov.pagopa.register.repository.initiative;
 
-import it.gov.pagopa.register.configuration.initiative.model.InitiativeConfig;
-import it.gov.pagopa.register.model.operation.Product;
-import it.gov.pagopa.register.repository.operation.ProductSpecificRepository;
+import it.gov.pagopa.register.model.initiative.InitiativeConfig;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
 
 
 @Repository
